refactor(deployment): move Nexus stack lifecycle to deployment
Slice 05 moves EnsureNexusStack behavior from the Nexus service package into the Deployment application service boundary while preserving the legacy Nexus import path as a compatibility facade.

Behavior is unchanged: the service still uses the compose repository and Portainer client ports to create or update the stack. Nexus readiness, admin access, anonymous access, and repository behavior remain outside Deployment.

Tests move stack lifecycle coverage to deployment, add a legacy import compatibility check, and add an architecture guard for Deployment ownership. arc42 and workflow metadata now include the stack lifecycle documentation scope.

Verification: git diff --check PASS; venv/bin/python tools/quality_gate.py quality PASS with ruff, import-linter, arch-tests, mypy, and unittest; focused unittest PASS with 10 tests.

Slice-ID: 05

workflowVersion: init-safety-boundary-separation-20260523

Rollback reference: revert this Slice 05 checkpoint commit
