+++ tests/application/services/nexus/test_bootstrap_nexus.py
@@ -4,42 +4,8 @@ from unittest.mock import MagicMock, patch
 from tiny_swarm_world.application.services.nexus.bootstrap_nexus import BootstrapNexus
 from tiny_swarm_world.application.services.nexus.enable_nexus_anonymous_access import EnableNexusAnonymousAccess
 from tiny_swarm_world.application.services.nexus.ensure_nexus_admin_access import EnsureNexusAdminAccess
-from tiny_swarm_world.application.services.nexus.ensure_nexus_stack import EnsureNexusStack
 from tiny_swarm_world.application.services.nexus.wait_for_nexus_ready import WaitForNexusReady
-from tiny_swarm_world.domain.deployment.stack_definition import StackDefinition
 from tiny_swarm_world.domain.nexus.nexus_user import NexusUser
-
-
-class TestEnsureNexusStack(unittest.TestCase):
-    def test_creates_stack_when_missing(self):
-        compose_repository = MagicMock()
-        compose_repository.get_compose_of.return_value = StackDefinition(name="nexus", compose_content="services: {}")
-        portainer_client = MagicMock()
-        portainer_client.get_endpoint_id_by_name.return_value = 1
-        portainer_client.find_stack_id_by_name.return_value = None
-
-        service = EnsureNexusStack(compose_repository, portainer_client, "nexus", "local")
-        service.run()
-
-        portainer_client.create_stack.assert_called_once()
-        portainer_client.update_stack.assert_not_called()
-
-    def test_updates_stack_when_present(self):
-        compose_repository = MagicMock()
-        compose_repository.get_compose_of.return_value = StackDefinition(name="nexus", compose_content="services: {}")
-        portainer_client = MagicMock()
-        portainer_client.get_endpoint_id_by_name.return_value = 1
-        portainer_client.find_stack_id_by_name.return_value = 7
-
-        service = EnsureNexusStack(compose_repository, portainer_client, "nexus", "local")
-        service.run()
-
-        portainer_client.update_stack.assert_called_once_with(
-            7,
-            StackDefinition(name="nexus", compose_content="services: {}"),
-            1,
-        )
-        portainer_client.create_stack.assert_not_called()
 
 
 class TestWaitForNexusReady(unittest.TestCase):
